multiple server endpoint inproved

+++ jsonGenerator/filesystem.js
@@ -1,4 +1,5 @@
 var fs = require('fs');
+var path = require('path');
 
 exports.saveJson2File = function(json, filePath) {
 	fs.writeFileSync(filePath, JSON.stringify(json, null, 4));
@@ -11,10 +12,11 @@ exports.cleanJSON = function(filesPath) {
 		function(file,index){
 			var curPath = filesPath + "/" + file;
 			if(!fs.lstatSync(curPath).isDirectory()) { 
-				fs.unlinkSync(curPath);
+				if(path.extname(file).toUpperCase() == "JSON")
+					fs.unlinkSync(curPath);
 			}
 		}
     );
 
-	console.log("--- Successfully deleted all files from '"+filesPath+"'");
+	console.log("--- Successfully deleted all JSON files from '"+filesPath+"'");
 }
